app/genemo-back/requirements.txt
absl-py==1.4.0
array-record==0.2.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dm-tree==0.1.8
etils==1.3.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
keras==2.12.0
keras-cv==0.4.0
libclang==15.0.6
linecache2==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
promise==2.3
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
